refactor: settings/diary/themeconfig preview 다듬기 + 파일 분리
- DiaryHistoryDialogs 를 별도 파일로 분리 + preview 추가.
- SettingsPage 의 컴포넌트들 (SettingItem, SettingColumn, NewBadge) 을
  SettingsComponents.kt 로 분리 (관련 preview 동반 이동).
- PersonalInformationPolicy/ServiceInfo/TeamInfo Screen preview 제거
  (WebView 라 preview 의미 없음, SimpleTopBar 만 보이면 redundant).
- PoorSwitch_On/Off preview 제거.
- ThemeDetailScreen_Success mock 에 themeDetailSampleLectures 적용
  (lectures 보강 + 색 조합 섹션이 의미 있게 표시).
