+++ app/src/main/java/com/wafflestudio/snutt2/feature/settings/TeamInfoPage.kt
@@ -17,8 +17,6 @@ import com.wafflestudio.snutt2.R
 import com.wafflestudio.snutt2.logging.AnalyticsScreen
 import com.wafflestudio.snutt2.logging.compose.logImpression
 import com.wafflestudio.snutt2.ui.components.compose.SimpleTopBar
-import com.wafflestudio.snutt2.ui.preview.SnuttPreview
-import com.wafflestudio.snutt2.ui.preview.SnuttPreviewSurface
 import com.wafflestudio.snutt2.ui.theme.ThemeMode
 
 @Composable
@@ -70,17 +68,6 @@ private fun TeamInfoScreen(
                     )
                 }
             },
-        )
-    }
-}
-
-@SnuttPreview
-@Composable
-private fun TeamInfoScreen_Default() {
-    SnuttPreviewSurface {
-        TeamInfoScreen(
-            themeMode = ThemeMode.AUTO,
-            onNavigateBack = {},
         )
     }
 }
